Fix: removed the button from products list and path products list is now homePage
