+++ src/App.js
@@ -7,7 +7,6 @@ import PerfilPage from "./components/PerfilPage.js"
 import GlobalStyle from "./styles/globalStyles.js"
 import { useState } from "react"
 import UserContext from "./contexts/UserContext.js"
-import ProductsListPage from "./components/ProductListPage.js"
 import KartContext from "./contexts/KartContext.js"
 
 function App() {
@@ -23,7 +22,6 @@ function App() {
             <Route path="/" element={<HomePage />} />
             <Route path="/login" element={<LoginPage />} />
             <Route path="/register" element={<RegisterPage />} />
-            <Route path="/home" element={<ProductsListPage />} />
             <Route path="/checkout" element={<CheckoutPage />} />
             <Route path="/perfil" element={<PerfilPage />} />
           </Routes>
